% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quack.R
\name{quack}
\alias{quack}
\title{Convert data collected with academictwitteR into DuckDB}
\usage{
quack(
  data_path,
  db = ":memory:",
  db_close = FALSE,
  convert_date = FALSE,
  return_con = TRUE,
  verbose = TRUE,
  error = FALSE
)
}
\arguments{
\item{data_path}{path to the directory hosting data collected with academictwitteR}

\item{db}{path to the DuckDB; the default value ":memory:" converts the data into a DuckDB in main memory}

\item{db_close}{whether to close the DuckDB after the operation. It is better to set it to TRUE, if \code{db} is not ":memory:"}

\item{convert_date}{whether or not \code{created_at} and \code{user_created_at} should be converted to date}

\item{return_con}{whether or not to return the connnection object; if false, the argument db is returned invisibly.}

\item{verbose}{whether to display a progress bar}

\item{error}{whether to raise error if there is no data}
}
\description{
This function coverts data collected with academictwitteR into DuckDB.
}
