% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quack.R
\name{quack}
\alias{quack}
\title{Convert data collected with academictwitteR into DuckDB}
\usage{
quack(data_path, db = ":memory:", db_close = FALSE, return_con = TRUE)
}
\arguments{
\item{data_path}{path to the directory hosting data collected with academictwitteR}

\item{db}{path to the DuckDB; the default value ":memory:" converts the data into a DuckDB in main memory}

\item{db_close}{whether to close the DuckDB after the operation. It is better to set it to TRUE, if \code{db} is not ":memory:"}

\item{return_con}{whether or not to return the connnection object}
}
\description{
This function coverts data collected with academictwitteR into DuckDB.
}
